CI, Only Build and Push Container on Main
This patch adds a conditional to the `Build and push Docker image` job so that this only runs on the `main` branch.
The indentation of the top level `on:` option has also been fixed.

Potential followups:
We might want to look into generating docker version tags dynamically since currently we push all merged changes to the same `0.1` tag.
If we introduce dynamically generated versions we technically could even use that to tag PRs if we really want that.