Refactor, Generalize SenderAdapter Code
This patch moves sender adapter specific code into a centralized place so that adding support for more sender adapters becomes easier.
This should aid us in adding support for all adapters supported by Swoosh.
The UI currently still needs to define views and "translations" for the adapter names explicitly to allow more freedom.

One side effect to this is that we can't reference `Keila.Mailings.Sender.Config` in the adapter code since that would introduce cyclic dependencies.

Currently the UI code needs to defined some methods to handle the test sender adapter. That's not great but I wonder if we even need to support a test specific adapter or if we can use the `SMTP` adapter in all cases?

This is meant to provide a starting point for a discussion so I'm not saying that this needs to be the final solution.