Concurrency, Prevent double sending of Campaigns
While trying to implement the deliver campaign later feature I noticed that we didn't properly guard against double sendings of campaigns.
The idea here is to use lock on the campaign and use `sent_at` as a guard to ensure that the campaign is not sended twice.

This came up since I tried to implement the send later feature for campaigns based on a new `deliver_at` field in combination with a Oban cron job rather than just scheduling jobs for each scheduled campaign.
The benefit of that is more control over sending of those scheduled campaigns and having the knowledge about the time the campaign will be delivered inside the domain rather than having it only inside Oban (or introducing a potential disconnect between those sources of truth).
Based on my experience this pattern is better scalable and adjustable for different needs than just scheduling jobs in the future.