Add ProseMirror-powered Markdown WYSIWYG  editor
This commit adds a new ProseMirror-powered WYSIWYG editor for Markdown
campaigns.

It also includes a small refactoring of the existing LiveView hook as
well as the addition of custom icons to go with the new editor.

The commit also fixes a bug in which saving campaigns would fail when no
sender was set.