Add UI for scheduling campaigns
This adds a Schedule button to the campaign edit view.
Error handling is somewhat rudimentary; an error message is displayed
when users try to perform an invalid scheduling action

This commit also adds an error message when trying to send a campaign
with no sender instead of a silent error.