Improve Markdown editing
- Functional preview button
- Recognition of Liquid tags
- No more escapting in Liquid tags

One thing that’s not quite solved yet is mixing liquid tags with
other markup.
Things will break if users try to make part of a liquid tag bold or
italic, but otherwise it wouldn’t be possible to have the entire output
of the tag rendered in bold print or italics.